+++ confidant/scripts/manage.py
@@ -6,6 +6,11 @@ from confidant.scripts.utils import RevokeGrants
 from confidant.scripts.utils import CreateDynamoTables
 from confidant.scripts.bootstrap import GenerateSecretsBootstrap
 from confidant.scripts.bootstrap import DecryptSecretsBootstrap
+from confidant.scripts.migrate import (
+    MigrateBlindCredentialSetAttribute,
+    MigrateServiceSetAttribute,
+)
+from confidant.scripts.migrate_bool import MigrateBooleanAttribute
 
 manager = Manager(app.app)
 
@@ -23,6 +28,14 @@ manager.add_command("decrypt_secrets_bootstrap", DecryptSecretsBootstrap)
 
 # Create dynamodb tables
 manager.add_command("create_dynamodb_tables", CreateDynamoTables)
+
+# Migration scripts
+manager.add_command("migrate_blind_cred_set_attribute",
+                    MigrateBlindCredentialSetAttribute)
+manager.add_command("migrate_service_set_attribute",
+                    MigrateServiceSetAttribute)
+manager.add_command("migrate_boolean_attribute",
+                    MigrateBooleanAttribute)
 
 
 def main():